backend/requirements.txt
fastapi[standard]
python-dateutil
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
